requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
requests==2.31.0
aiofiles==23.2.0
pydantic==2.5.0
email-validator==2.1.0
fastapi-cors==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

